+++ saleor/graphql/api.py
@@ -3,7 +3,7 @@ from graphene_federation import build_schema
 from .account.schema import AccountMutations, AccountQueries
 from .app.schema import AppMutations, AppQueries
 from .checkout.schema import CheckoutMutations, CheckoutQueries
-from .core.schema import CoreMutations, CoreQueries
+from .core.schema import CoreQueries
 from .csv.schema import CsvMutations, CsvQueries
 from .discount.schema import DiscountMutations, DiscountQueries
 from .giftcard.schema import GiftCardMutations, GiftCardQueries
@@ -49,7 +49,6 @@ class Mutation(
     AccountMutations,
     AppMutations,
     CheckoutMutations,
-    CoreMutations,
     CsvMutations,
     DiscountMutations,
     PluginsMutations,