+++ CHANGELOG.md
@@ -10,6 +10,9 @@ All notable, unreleased changes to this project will be documented in this file.
 - Add products csv export - #5255 by @IKarbowiak
 - Explicit country assignment in default shipping zones - #5736 by @maarcingebala
 - Drop `json_content` field from the `Menu` model - #5761 by @maarcingebala
+- Refactor JWT support - #5734 by @korycins
+- Strip warehouse name in mutations - #5766 by @koradon
+- Add missing OrderEvents during checkout flow - #5684 by @koradon
 
 ## 2.10.1